+++ daemon/tests/ipc_integration.rs
@@ -55,6 +55,8 @@ fn test_app_state() -> Arc<AppState> {
         rate_c_per_s: None,
         session_min_c: None,
         session_max_c: None,
+        chip_name: "k10temp".into(),
+        temp_type: None,
     }]);
 
     Arc::new(AppState {